REF move code to _tables.py; cleanup
