+++ civis/io/__init__.py
@@ -1,9 +1,8 @@
 from ._databases import query_civis, transfer_table
 from ._files import file_to_civis, civis_to_file
-from ._s3 import civis_multifile_unload
 from ._tables import (read_civis, read_civis_sql, civis_to_csv,
-                      dataframe_to_civis, csv_to_civis)
+                      civis_to_multifile_csv, dataframe_to_civis, csv_to_civis)
 
 __all__ = ["query_civis", "transfer_table", "file_to_civis", "civis_to_file",
            "read_civis", "read_civis_sql", "civis_to_csv",
-           "dataframe_to_civis", "csv_to_civis", "civis_multifile_unload"]
+           "dataframe_to_civis", "csv_to_civis", "civis_to_multifile_csv"]
